Uplift Python API docs
This continues the work done for #613. Now all relevant symbols
are directly documented, for every top-level package.

In doing so, I added a first set of markers to shape the API
streamlining that will come with v2.0.
Ping #624
Ping #616

Some implementations that continued to be located in `__init__.py`
files have been move into new, dedicated source files.
