+++ datalad_next/consts/__init__.py
@@ -1,4 +1,19 @@
 """Common constants
+
+COPY_BUFSIZE
+  ``shutil`` buffer size default, with Windows platform default changes
+  backported from Python 3.10.
+
+PRE_INIT_COMMIT_SHA
+  SHA value for ``git hash-object -t tree /dev/null``, i.e. for nothing.
+  This corresponds to the state of a Git repository before the first commit
+  is made.
+
+on_linux
+  ``True`` if executed on the Linux platform.
+
+on_windows
+  ``True`` if executed on the Windows platform.
 """
 
 # import from "utils", but these really are constants
