+++ datalad_next/exceptions/__init__.py
@@ -1,4 +1,14 @@
-"""All custom exceptions used in datalad-next"""
+"""Special purpose exceptions
+
+.. currentmodule:: datalad_next.exceptions
+.. autosummary::
+   :toctree: generated
+
+   CapturedException
+   IncompleteResultsError
+   NoDatasetFound
+"""
+# we cannot have CommandError above, sphinx complains
 
 # TODO rethink the purpose of this module and possibly
 # make it about *external* custom exceptions
@@ -9,6 +19,7 @@ from datalad.support.exceptions import (
     NoDatasetFound,
 )
 
+# TODO REMOVE FOR V2.0 (they are specific to that module
 from datalad_next.url_operations import (
     UrlOperationsRemoteError,
     UrlOperationsAuthenticationError,
